sl-49: * moved updateUser action to user slice
