+++ frontend/src/slices/auth/actions.ts
@@ -5,7 +5,6 @@ import { type AsyncThunkConfig } from '~/libs/types/types.js';
 import {
   type UserAuthResponse,
   type UserSignUpRequestDto,
-  type UserUpdateRequestDto,
 } from '~/packages/users/users.js';
 
 import { name as sliceName } from './auth.slice.js';
@@ -23,17 +22,4 @@ const signUp = createAsyncThunk<
   return user;
 });
 
-const updateUser = createAsyncThunk<
-  UserAuthResponse,
-  UserUpdateRequestDto,
-  AsyncThunkConfig
->(
-  `${sliceName}/update-user-details`,
-  async (updateUserDetailsPayload, { extra }) => {
-    const { userApi } = extra;
-
-    return await userApi.updateUserDetails(updateUserDetailsPayload);
-  },
-);
-
-export { signUp, updateUser };
+export { signUp };
