+++ frontend/src/slices/auth/auth.ts
@@ -1,10 +1,9 @@
-import { signUp, updateUser } from './actions.js';
+import { signUp } from './actions.js';
 import { actions } from './auth.slice.js';
 
 const allActions = {
   ...actions,
   signUp,
-  updateUser,
 };
 
 export { allActions as actions };
