+++ frontend/src/slices/users/actions.ts
@@ -1,7 +1,11 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { type AsyncThunkConfig } from '~/libs/types/types.js';
-import { type UserGetAllResponseDto } from '~/packages/users/users.js';
+import {
+  type UserAuthResponse,
+  type UserGetAllResponseDto,
+  type UserUpdateRequestDto,
+} from '~/packages/users/users.js';
 
 import { name as sliceName } from './users.slice.js';
 
@@ -15,4 +19,14 @@ const loadAll = createAsyncThunk<
   return userApi.getAll();
 });
 
-export { loadAll };
+const updateUser = createAsyncThunk<
+  UserAuthResponse,
+  UserUpdateRequestDto,
+  AsyncThunkConfig
+>(`${sliceName}/update-user`, async (updateUserPayload, { extra }) => {
+  const { userApi } = extra;
+
+  return await userApi.updateUser(updateUserPayload);
+});
+
+export { loadAll, updateUser };
