+++ frontend/src/slices/users/users.ts
@@ -1,9 +1,10 @@
-import { loadAll } from './actions.js';
+import { loadAll, updateUser } from './actions.js';
 import { actions } from './users.slice.js';
 
 const allActions = {
   ...actions,
   loadAll,
+  updateUser,
 };
 
 export { allActions as actions };
